вынес конструктор начинки для пиццы в отдельный компонент
