+++ src/frontend/src/views/Index.vue
@@ -12,44 +12,10 @@
             <BuilderSizeSelector :sizes="sizes" />
           </div>
           <div class="content__ingredients">
-            <div class="sheet">
-              <h2 class="title title--small sheet__title">
-                Выберите ингредиенты
-              </h2>
-              <div class="sheet__content ingredients">
-                <div class="ingredients__sauce">
-                  <p>Основной соус:</p>
-                  <RadioButton
-                    v-for="sauce in sauces"
-                    :key="sauce.id"
-                    name="sauce"
-                    :value="sauce.name"
-                    :checked="sauce.id === 1"
-                    class="ingredients__input"
-                  >
-                    <span>{{ sauce.name }}</span>
-                  </RadioButton>
-                </div>
-                <div class="ingredients__filling">
-                  <p>Начинка:</p>
-                  <ul class="ingredients__list">
-                    <li
-                      v-for="ingredient in ingredients"
-                      :key="ingredient.id"
-                      class="ingredients__item"
-                    >
-                      <span
-                        class="filling"
-                        :class="setFillingLabelClass(ingredient.name)"
-                      >
-                        {{ ingredient.name }}
-                      </span>
-                      <ItemCounter class="ingredients__counter" />
-                    </li>
-                  </ul>
-                </div>
-              </div>
-            </div>
+            <BuilderIngredientsSelector
+              :ingredients="ingredients"
+              :sauces="sauces"
+            />
           </div>
           <div class="content__pizza">
             <label class="input">
@@ -84,20 +50,18 @@
 <script>
 import pizza from "@/static/pizza.json";
 import AppLayout from "@/layouts/AppLayout.vue";
-import RadioButton from "@/common/components/RadioButton.vue";
-import ItemCounter from "@/common/components/ItemCounter.vue";
 import BuilderDoughSelector from "@/modules/builder/components/BuilderDoughSelector.vue";
 import BuilderSizeSelector from "@/modules/builder/components/BuilderSizeSelector.vue";
+import BuilderIngredientsSelector from "@/modules/builder/components/BuilderIngredientsSelector.vue";
 
 export default {
   name: "Index",
 
   components: {
     AppLayout,
-    RadioButton,
-    ItemCounter,
     BuilderDoughSelector,
     BuilderSizeSelector,
+    BuilderIngredientsSelector,
   },
 
   data() {
@@ -107,29 +71,6 @@ export default {
       sizes: pizza.sizes,
       dough: pizza.dough,
     };
-  },
-
-  methods: {
-    setFillingLabelClass(name) {
-      const ingredients = {
-        Грибы: "mushrooms",
-        Томаты: "tomatoes",
-        Ананас: "ananas",
-        Бекон: "bacon",
-        "Блю чиз": "blue_cheese",
-        Чеддер: "cheddar",
-        Чили: "chile",
-        Ветчина: "ham",
-        Халапеньо: "jalapeno",
-        Моцарелла: "mozzarella",
-        Маслины: "olives",
-        Лук: "onion",
-        Пармезан: "parmesan",
-        Салями: "salami",
-        Лосось: "salmon",
-      };
-      return `filling--${ingredients[name]}`;
-    },
   },
 };
 </script>
